Update robb-smarrt-8-switch-ROB_200-007-0.yaml
Thanks for the blueprint first off, very helpfull!

The long press of the off buttons is not working. The event that is fired is 'move' not 'move_with_on_off'. I made changes to include that test and also test for 'mode == MoveMode.Down'. It works for me now but probably the community may benefit from this as well.

Perhaps the events fired by the Robb buttons changed causing the test to break?